pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jobmaster-backend"
version = "0.1.0"
description = "Backend MVP for generating tailored resumes from CV, career dump, and JD"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0",
  "uvicorn[standard]>=0.35.0",
  "python-multipart>=0.0.20",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.0",
  "google-generativeai>=0.8.5",
  "openai>=1.99.0",
]

[tool.setuptools]
package-dir = {"" = "backend"}

[tool.setuptools.packages.find]
where = ["backend"]
